pyproject.toml
[project]
name = "ai-log-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
